Enum haybale_pitchfork::CompleteAbstractData

pub enum CompleteAbstractData {
    PublicValue {
        bits: usize,
        value: AbstractValue,
    },
    Secret {
        bits: usize,
    },
    Array {
        element_type: Box<Self>,
        num_elements: usize,
    },
    Struct {
        name: String,
        elements: Vec<Self>,
    },
    PublicPointerTo {
        pointee: Box<Self>,
        maybe_null: bool,
    },
    PublicPointerToFunction(String),
    PublicPointerToHook(String),
    PublicPointerToSelf,
    PublicPointerToParentOr(Option<Box<CompleteAbstractData>>),
    VoidOverride {
        llvm_struct_name: Option<String>,
        data: Box<Self>,
    },
    SameSizeOverride {
        data: Box<Self>,
    },
    WithWatchpoint {
        name: String,
        data: Box<Self>,
    },
}

An abstract description of a value: if it is public or not, if it is a pointer or not, does it point to data that is public/secret, maybe it's a struct with some public and some secret fields, etc.

Unlike AbstractData, these may never be "underspecified" - that is, they must be a complete description of the data structure.

Variants

PublicValue

A public value, of the given size in bits. The AbstractValue is used to indicate whether the value should have a particular concrete value, be unconstrained, etc.

This may be used for either a non-pointer value, or for a pointer value if you want to specify the exact numerical value of the pointer (e.g. NULL).

Fields of PublicValue

bits: usizevalue: AbstractValue

Secret

A secret value (pointer or non-pointer, doesn't matter) of the given size in bits

Fields of Secret

bits: usize

Array

A (first-class) array of values

Fields of Array

element_type: Box<Self>num_elements: usize

Struct

A (first-class) structure of values

Fields of Struct

name: Stringelements: Vec<Self>

PublicPointerTo

A (public) pointer to something - another value, an array, etc

Fields of PublicPointerTo

pointee: Box<Self>

Description of the thing being pointed to

maybe_null: bool

If false, the pointer must point to the pointee; if true, it may either point to the pointee or be NULL

PublicPointerToFunction(String)

A (public) pointer to the LLVM Function with the given name

PublicPointerToHook(String)

A (public) pointer to the hook registered for the given name

PublicPointerToSelf

A (public) pointer to this struct itself. E.g., in the C code

struct Foo {
    int x;
    Foo* f;
};

you could use this for Foo* f to indicate it should point to this exact Foo itself.

PublicPointerToParentOr(Option<Box<CompleteAbstractData>>)

A (public) pointer to this struct's parent. E.g., in the C code

struct Foo {
    int x;
    Bar* bar1;
    Bar* bar2;
    ...
};

struct Bar {
    int y;
    Foo* parent;  // pointer to the Foo containing this Bar
};

you could use this for Foo* parent to indicate it should point to the Foo containing this Bar.

If the Option is Some, then if the parent is not the correct type (or if there is no parent, i.e., we are directly initializing this) then pointer to the given CompleteAbstractData instead

VoidOverride

When C code uses void*, this often becomes i8* in LLVM. However, within Pitchfork, we may want to specify some type other than i8* for the purposes of allocating and analyzing the data behind the void*.

This says to use the provided CompleteAbstractData even though the LLVM type is i8.

If the optional llvm_struct_name is included, it will lookup that struct's type and check against that. Otherwise, no typechecking will be performed and the provided CompleteAbstractData will be assumed correct.

Fields of VoidOverride

llvm_struct_name: Option<String>data: Box<Self>

SameSizeOverride

Use the given data, even though it may not match the LLVM type. It still needs to be the same size (number of bits) as the LLVM type. For instance, you could specify that some LLVM pointer-size integer should actually be initialized to have a pointer value and point to some specified data.

To override a void* type, see VoidOverride - and this probably won't work for that anyways because of the same-size restriction. See comments on VoidOverride.

Fields of SameSizeOverride

data: Box<Self>

WithWatchpoint

Use the given data, but also (during initialization) add a watchpoint with the given name to the State covering the memory region it occupies.

Fields of WithWatchpoint

name: Stringdata: Box<Self>

Methods

impl CompleteAbstractData

pub fn pub_i8(value: AbstractValue) -> Self

an 8-bit public value

pub fn pub_i16(value: AbstractValue) -> Self

a 16-bit public value

pub fn pub_i32(value: AbstractValue) -> Self

a 32-bit public value

pub fn pub_i64(value: AbstractValue) -> Self

a 64-bit public value

pub fn pub_integer(bits: usize, value: AbstractValue) -> Self

a public value with the given number of bits

pub fn sec_i8() -> Self

an 8-bit secret value

pub fn sec_i16() -> Self

a 16-bit secret value

pub fn sec_i32() -> Self

a 32-bit secret value

pub fn sec_i64() -> Self

a 64-bit secret value

pub fn sec_integer(bits: usize) -> Self

a secret value with the given number of bits

pub fn pub_pointer_to(data: Self) -> Self

a (public) pointer to something - another value, an array, etc

pub fn pub_maybe_null_pointer_to(data: Self) -> Self

A (public) pointer which may either point to the given data or be NULL

pub fn pub_pointer_to_func(funcname: impl Into<String>) -> Self

a (public) pointer to the LLVM Function with the given name

pub fn pub_pointer_to_hook(funcname: impl Into<String>) -> Self

a (public) pointer to the hook registered for the given name

pub fn pub_pointer_to_self() -> Self

a (public) pointer to this struct itself; see comments on CompleteAbstractData::PublicPointerToSelf

pub fn pub_pointer_to_parent() -> Self

A (public) pointer to this struct's parent. E.g., in the C code

struct Foo {
    int x;
    Bar* bar1;
    Bar* bar2;
    ...
};

struct Bar {
    int y;
    Foo* parent;  // pointer to the Foo containing this Bar
};

you could use this for Foo* parent to indicate it should point to the Foo containing this Bar.

pub fn pub_pointer_to_parent_or(data: Self) -> Self

Like pub_pointer_to_parent(), but if the parent is not the correct type (or if there is no parent, i.e., we are directly initializing this) then pointer to the given CompleteAbstractData instead

pub fn array_of(element_type: Self, num_elements: usize) -> Self

A (first-class) array of values

pub fn _struct(
name: impl Into<String>,
elements: impl IntoIterator<Item = Self>
) -> Self

A (first-class) structure of values. Name used only for debugging purposes, need not match the (mangled) LLVM struct name.

(_struct used instead of struct to avoid collision with the Rust keyword)

pub fn unconstrained_pointer() -> Self

A (public) pointer which may point anywhere, including being NULL

pub fn void_override(llvm_struct_name: Option<&str>, data: Self) -> Self

When C code uses void*, this often becomes i8* in LLVM. However, within Pitchfork, we may want to specify some type other than i8* for the purposes of allocating and analyzing the data behind the void*.

This says to use the provided CompleteAbstractData even though the LLVM type is i8.

If the optional llvm_struct_name is included, it will lookup that struct's type and check against that. Otherwise, no typechecking will be performed and the provided CompleteAbstractData will be assumed correct.

pub fn same_size_override(data: Self) -> Self

Use the given data, even though it may not match the LLVM type. It still needs to be the same size (number of bits) as the LLVM type. For instance, you could specify that some LLVM pointer-size integer should actually be initialized to have a pointer value and point to some specified data.

To override a void* type, see void_override - and this probably won't work for that anyways because of the same-size restriction. See comments on void_override.

pub fn with_watchpoint(name: impl Into<String>, data: Self) -> Self

Use the given data, but also (during initialization) add a watchpoint with the given name to the State covering the memory region it occupies.

impl CompleteAbstractData

pub const POINTER_SIZE_BITS: usize

pub fn size_in_bits(&self) -> usize

Get the size of the CompleteAbstractData, in bits

pub fn field_size_in_bits(&self, n: usize) -> usize

Get the size of the nth (0-indexed) field/element of the CompleteAbstractData, in bits. The CompleteAbstractData must be a Struct or Array.

pub fn offset_in_bits(&self, n: usize) -> usize

Get the offset of the nth (0-indexed) field/element of the CompleteAbstractData, in bits. The CompleteAbstractData must be a Struct or Array.

pub fn is_pointer(&self) -> bool

Does the CompleteAbstractData represent a pointer of some kind?

pub fn pointee_size_in_bits(&self) -> usize

Get the size of the data this CompleteAbstractData points to.

Panics if self is not a pointer of some kind.