Function haybale_pitchfork::is_constant_time_in_inputs

pub fn is_constant_time_in_inputs<'p>(
    funcname: &'p str, 
    project: &'p Project, 
    config: Config<'p, Backend>
) -> bool

Is a function "constant-time" in its inputs. That is, does the function ever make branching decisions, or perform address calculations, based on its inputs.

For argument descriptions, see haybale::symex_function().