Module haybale_pitchfork::hook_helpers

This module contains helper functions that may be useful in writing function hooks.

Functions

allocate_and_init_abstractdata
Allocates space for the given AbstractData, initializes it, and returns a pointer to the newly-allocated space.

fill_secret_with_length
Fills a buffer with secret data, and also outputs the number of bytes written.

fill_unconstrained_with_length
Fills a buffer with unconstrained data, and also outputs the number of bytes written.

reinitialize_pointee
Reinitializes whatever is pointed to by the given pointer, according to the given AbstractData.