Module haybale_pitchfork::hooks

This module contains a few simple built-in function hooks which can be used with Config.function_hooks.

Functions

propagate_taint
This hook will return a secret value if any of the arguments are secret, or if any of the arguments contain a pointer to any secret data. Otherwise, it will return an unconstrained public value.

return_public_unconstrained
This hook will ignore all of the function arguments and simply return an unconstrained public value of the appropriate size, or void for void-typed functions.

return_secret
This hook will ignore all of the function arguments and simply return a secret value of the appropriate size, or void for void-typed functions.