Function haybale_pitchfork::hooks::propagate_taint

pub fn propagate_taint(
    proj: &Project, 
    state: &mut State<Backend>, 
    call: &dyn IsCall
) -> Result<ReturnValue<BV>>

This hook will return a secret value if any of the arguments are secret, or if any of the arguments contain a pointer to any secret data. Otherwise, it will return an unconstrained public value.

Assumes that opaque struct types do not contain secret data or pointers to secret data.