Module haybale_pitchfork::secret

This module contains the dynamic taint-tracking layer implemented on top of haybale. It provides a haybale::backend::Backend which performs dynamic taint tracking and reports constant-time violations.

The BV, Memory, and Backend in this module are intended to be used qualified whenever there is a chance of confusing them with haybale::backend::{BV, Memory, Backend}, haybale::{memory,simple_memory}::Memory, or boolector::BV.

Structs

Backend
A Backend which performs dynamic taint tracking and reports constant-time violations.

BtorRef
This wrapper around Rc<Btor> exists simply so we can give it a different implementation of haybale::backend::SolverRef than the one provided by haybale::backend.

Memory
A Memory which tracks which of its contents are public or secret, and reports constant-time violations whenever secret data is used as an address for operations on it.

Enums

BV
A wrapper around boolector::BV which can represent either public or secret data.