Enum haybale_pitchfork::secret::BV

pub enum BV {
    Public(BV<Rc<Btor>>),
    Secret {
        btor: BtorRef,
        width: u32,
        symbol: Option<String>,
    },
    PartiallySecret {
        secret_mask: Vec<bool>,
        data: BV<Rc<Btor>>,
        symbol: Option<String>,
    },
}

A wrapper around boolector::BV which can represent either public or secret data.

Variants

Public(BV<Rc<Btor>>)

Secret

Secret values are opaque because we don't care about their actual value, only how they are used.

Fields of Secret

btor: BtorRefwidth: u32symbol: Option<String>

PartiallySecret

PartiallySecret values have some secret and some not-secret bits. Currently we make just a best-effort attempt to handle these; for now, some functions may just fall back on treating the entire value as Secret.

Fields of PartiallySecret

secret_mask: Vec<bool>

A vector the length of the PartiallySecret value's bitwidth, where each true indicates that the corresponding position in the PartiallySecret is secret, and likewise false indicates public. secret_mask[0] corresponds to the bit at position 0, which is the rightmost (least-significant) bit.

data: BV<Rc<Btor>>

A BV, which must have bitwidth exactly equal to the length of secret_mask. In positions where secret_mask is false, this gives the value of the public bits. In positions where secret_mask is true, the data here is invalid and should not be used for anything.

symbol: Option<String>

Methods

impl BV

pub fn is_secret(&self) -> bool

pub fn as_public(&self) -> &BV<Rc<Btor>>

Gets the value out of a BV::Public, panicking if it is instead a BV::Secret